# src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "REST API",
        "description": "Decumentação da api",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:8080/trade",
    "basePath": "/",
    "tags": [{
        "name": "User",
        "description": "Api for users in the system"
    }],
    "schemes": [
        "http"
    ],
    "consumes": [],
    "produces": [],
    "paths": {
        "/lista": {
            "get": {
                "tags": [
                    "Stocks"
                ],
                "description": "",
                "parameters": [],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/stocks-favoritos": {
            "get": {
                "tags": [
                    "Stocks"
                ],
                "description": "",
                "parameters": [],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/set-favorite/{id_trade}": {
            "put": {
                "tags": [
                    "Stocks"
                ],
                "description": "",
                "parameters": [{
                        "name": "id_trade",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "obj",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "favorito": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/login": {
            "parameters": [{
                "in": "body",
                "name": "body"
            }],
            "post": {
                "tags": [
                    "User"
                ],
                "description": "",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        },
        "/register": {
            "parameters": [{
                "in": "body",
                "name": "body",
                "schema": {
                    "$ref": "#/definitions/User"
                }
            }],
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Cadastra usuário",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        },
        "/update-user/{id_user}": {
            "put": {
                "tags": [
                    "User"
                ],
                "description": "",
                "parameters": [{
                        "name": "id_user",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "obj",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "favorito": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "definitions": {
        "User": {
            "properties": {
                "name": {
                    "type": "string",
                    "required": true
                },
                "email": {
                    "type": "string",
                    "required": true
                },
                "password": {
                    "type": "string",
                    "required": true
                }
            }
        },
        "Stocks": {
            "properties": {
                "stock": {
                    "type": "string",
                    "equired": true
                },
                "exchange": {
                    "type": "string",
                    "equired": true
                },
                "country": {
                    "type": "string",
                    "equired": true
                },
                "company": {
                    "type": "string",
                    "equired": true
                },
                "variation": {
                    "type": "number",
                    "equired": true
                },
                "price": {
                    "type": "number",
                    "equired": true
                },
                "favorito": {
                    "type": "boolean",
                    "default": false
                },
                "chart": {
                    "type": "array"
                }
            }
        }
    }
}